Restructure Twitch API package
